cmd/kube-client-agent: add new option csrName to specify the csr object's name
This change should not change any existing behaviour. If the field '--csrName' is
not specified then the commonname will be used for the object's name (what happens today).
When specified, the objects name will be taken accordingly. This is useful
when multiple CSR objects need to be created for the same 'commonname'.