*: add secure and insecure health checks
Since the kube-etcd-signer-server mimics the kube-apiserver's CSR apis to sign certs for etcd, it is usually run on the same port as the kube-apiserver.
In those cases serving health checks becomes important as kube-apiserver is usually behind the load-balancers with health checks.

Adds `/readyz` health check to the secure server.
also adds a flag to serve the same health check on an insecure transport as some cloud only support http health checks.